+++ game-frontend/projects/game-admin/src/app/components/upgrade-crud/upgrade-crud.component.ts
@@ -1,8 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { AdminUpgradeService } from '../../services/admin-upgrade.service';
 import { Upgrade, UpgradeType } from '@owge/universe';
+import { WidgetFilter, WidgetFilterUtil } from '@owge/widgets';
+import { Observable, Subject } from 'rxjs';
+import { AdminFactionService } from '../../services/admin-faction.service';
 import { AdminUpgradeTypeService } from '../../services/admin-upgrade-type.service';
-
+import { AdminUpgradeService } from '../../services/admin-upgrade.service';
 
 /**
  *
@@ -18,13 +20,32 @@ import { AdminUpgradeTypeService } from '../../services/admin-upgrade-type.servi
 })
 export class UpgradeCrudComponent implements OnInit {
 
+  public upgrades: Upgrade[];
+  public elsObservable: Observable<Upgrade[]>;
   public selectedEl: Upgrade;
   public upgradeTypes: UpgradeType[];
+  public secondValueFilters: WidgetFilter<any>[] = [WidgetFilterUtil.buildByNameFilter()];
 
-  constructor(public adminUpgradeService: AdminUpgradeService, private _adminUpgradeTypeService: AdminUpgradeTypeService) { }
+  private subject: Subject<Upgrade[]> = new Subject;
+
+  constructor(
+    public adminUpgradeService: AdminUpgradeService,
+    private _adminUpgradeTypeService: AdminUpgradeTypeService,
+    private adminFactionService: AdminFactionService
+  ) { }
 
   ngOnInit() {
     this._adminUpgradeTypeService.findAll().subscribe(val => this.upgradeTypes = val);
+    this.adminFactionService.buildFilter().then(result => this.secondValueFilters.push(result));
+    this.elsObservable = this.subject.asObservable();
+    this.adminUpgradeService.findAll().subscribe(upgradesResult => {
+      this.subject.next(upgradesResult);
+      this.upgrades = upgradesResult;
+    });
+  }
+
+  public onFilter(filtered: Upgrade[]) {
+    this.subject.next(filtered);
   }
 
 }