+++ game-frontend/projects/game-admin/src/app/services/admin-special-location.service.ts
@@ -1,13 +1,10 @@
 import { Injectable } from '@angular/core';
+import { ProgrammingError, RequirementInformation, validContext } from '@owge/core';
 import {
-    AbstractCrudService,
-    SpecialLocation,
-    WithImprovementsCrudMixin,
-    UniverseGameService,
-    CrudServiceAuthControl,
-    CrudConfig
+    AbstractCrudService, CrudConfig, CrudServiceAuthControl, SpecialLocation, UniverseGameService, WithImprovementsCrudMixin
 } from '@owge/universe';
-import { validContext } from '@owge/core';
+import { WidgetFilter } from '@owge/widgets';
+import { take } from 'rxjs/operators';
 import { Mixin } from 'ts-mixer';
 
 export interface AdminSpecialLocationService
@@ -28,6 +25,22 @@ export class AdminSpecialLocationService extends AbstractCrudService<SpecialLoca
     public constructor(protected _universeGameService: UniverseGameService) {
         super(_universeGameService);
         this._crudConfig = this.getCrudConfig();
+    }
+
+    public async buildFilter(): Promise<WidgetFilter<SpecialLocation>> {
+        return {
+            name: 'FILTER.BY_SPECIAL_LOCATION',
+            data: await this.findAll().pipe(take(1)).toPromise(),
+            filterAction: async (input, selectedSpecialLocation) => {
+                const requirements: RequirementInformation[] = input.requirements;
+                if (!requirements) {
+                    throw new ProgrammingError('Can NOT filter when the input has not requirements');
+                }
+                return requirements.some(requirement =>
+                    requirement.requirement.code === 'HAVE_SPECIAL_LOCATION' && requirement.secondValue === selectedSpecialLocation.id
+                );
+            }
+        };
     }
 
     protected _getEntity(): string {